requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
ezdxf==1.0.1
svgpathtools==1.6.0
lxml==5.1.0
pyclipper2==1.0.6
shapely==2.0.2
rtree==1.2.0
reportlab==4.1.0
Pillow==10.2.0
numpy==1.26.3
scipy==1.12.0
aiosqlite==0.19.0
pydantic==2.5.0
python-multipart==0.0.6
pytest==7.4.4
pytest-asyncio==0.23.2
